+++ spec/www/spec/db-tx-string-test.js
@@ -43,7 +43,7 @@ var mytests = function() {
         if (isWebSql) {
           return window.openDatabase(name, "1.0", "Demo", DEFAULT_SIZE);
         } else {
-          return window.sqlitePlugin.openDatabase({name: name, location: 'default'});
+          return window.sqlitePlugin.openDatabase({name: name});
         }
       }