+++ spec/www/spec/db-tx-value-bindings-test.js
@@ -71,7 +71,7 @@ var mytests = function() {
         if (isWebSql) {
           return window.openDatabase(name, "1.0", "Demo", DEFAULT_SIZE);
         } else {
-          return window.sqlitePlugin.openDatabase({name: name, location: 0});
+          return window.sqlitePlugin.openDatabase({name: name});
         }
       }